+++ include/render/primitives/BVH.h
@@ -88,6 +88,8 @@ namespace render {
     RayPacketIntersection4 intersectPacket(const Ray4& rays, render::State& state) const override;
     PrimitivePacketHit4 intersectPacketHits(const Ray4& rays,
                                             const PrimitivePacketState4& states) const override;
+    PrimitivePacketHit8 intersectPacketHits(const Ray8& rays,
+                                            const PrimitivePacketState8& states) const override;
 
 #if RAYTRACER_SIMD_AVX
     /**
@@ -153,6 +155,10 @@ namespace render {
                                 std::array<double, Ray4::lanes>& minDistances,
                                 PrimitivePacketHit4& result,
                                 const PrimitivePacketState4& states) const;
+    void intersectPacketHitNode(const Node* node, const Ray8& rays, uint16_t activeMask,
+                                std::array<double, Ray8::lanes>& minDistances,
+                                PrimitivePacketHit8& result,
+                                const PrimitivePacketState8& states) const;
 #if RAYTRACER_SIMD_AVX
     void intersectPacketNode(const Node* node, const Ray8& rays, uint16_t activeMask,
                              std::array<float, Ray8::lanes>& tMin, uint16_t& hitMask,